pyproject.toml
[tool.poetry]
name = "remcord-api-testing"
version = "0.1.0"
description = ""
authors = ["Nathan Solis <66754842+coldenate@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "remcord_api_testing"}]

[tool.poetry.dependencies]
python = "^3.11"
requests-oauthlib = "^1.3.1"
pyperclip = "^1.8.2"
python-dotenv = "^1.0.0"
fastapi = "^0.97.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
itsdangerous = "^2.1.2"
pydantic = "^1.10.9"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
